add select batch table
