+++ client/packages/system/src/Vaccination/Components/SelectBatch.tsx
@@ -1,0 +1,107 @@
+import {
+  BasicSpinner,
+  Checkbox,
+  Column,
+  ColumnAlign,
+  createTableStore,
+  DataTable,
+  NumUtils,
+  TableProvider,
+  useColumns,
+  useRowStyle,
+  useTranslation,
+} from '@openmsupply-client/common';
+import React, { useEffect } from 'react';
+import { StockLineFragment, useStockLines } from '../../Item';
+import { VaccinationStockLine } from '../api';
+
+interface SelectBatchProps {
+  itemId: string;
+  stockLine: VaccinationStockLine | null;
+  setStockLine: (stockLine: VaccinationStockLine) => void;
+}
+
+export const SelectBatch = ({
+  itemId,
+  stockLine,
+  setStockLine,
+}: SelectBatchProps) => {
+  const { data, isLoading } = useStockLines(itemId);
+
+  const columns = useColumns<StockLineFragment>(
+    [
+      {
+        key: 'select',
+        Cell: ({ rowData }) => (
+          <Checkbox
+            checked={rowData.id === stockLine?.id}
+            onClick={() => setStockLine(rowData)}
+          />
+        ),
+      },
+      'batch',
+      ['expiryDate', { align: ColumnAlign.Left }],
+      {
+        key: 'doses',
+        label: 'label.doses',
+        accessor: ({ rowData }) =>
+          NumUtils.round(
+            rowData.item.doses *
+              rowData.availableNumberOfPacks *
+              rowData.packSize
+          ),
+      },
+    ],
+    {},
+    [itemId, stockLine]
+  );
+
+  return (
+    <TableProvider createStore={createTableStore}>
+      {isLoading ? (
+        <BasicSpinner />
+      ) : (
+        <BatchTable
+          columns={columns}
+          data={data?.nodes ?? []}
+          setStockLine={setStockLine}
+        />
+      )}
+    </TableProvider>
+  );
+};
+
+const BatchTable = ({
+  columns,
+  data,
+  setStockLine,
+}: {
+  columns: Column<StockLineFragment>[];
+  data: StockLineFragment[];
+  setStockLine: (stockLine: VaccinationStockLine) => void;
+}) => {
+  const t = useTranslation('dispensary');
+  const { setRowStyles } = useRowStyle();
+
+  useEffect(() => {
+    setRowStyles(
+      data.map(r => r.id),
+      {
+        '& td': {
+          padding: 0,
+        },
+      }
+    );
+  }, [data]);
+
+  return (
+    <DataTable
+      id="vaccination-batches"
+      columns={columns}
+      data={data}
+      noDataMessage={t('messages.no-stock-available')}
+      onRowClick={row => setStockLine(row)}
+      dense
+    />
+  );
+};
